Refactoring DonationFinishedController
